moved tactical map to its own tier
