+++ client/package.json
@@ -7,6 +7,7 @@
     "@chakra-ui/icons": "^1.0.13",
     "@chakra-ui/react": "^1.6.1",
     "@cohdex/shared": "1.0.0",
+    "@cohdex/tactical-map": "1.0.0",
     "@emotion/react": "^11",
     "@emotion/styled": "^11",
     "@fontsource/karla": "^4.3.0",
