+++ tactical-map/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@cohdex/tactical-map",
+  "version": "1.0.0",
+  "types": "dist/index.d.ts",
+  "author": {
+    "email": "iamchets@yahoo.com",
+    "name": "Donny Roufs",
+    "url": "https://cohdex.com/"
+  },
+  "license": "MIT",
+  "main": "dist/index.js",
+  "scripts": {
+    "watch": "tsc -w"
+  },
+  "files": [
+    "dist"
+  ],
+  "dependencies": {
+    "@cohdex/shared": "1.0.0"
+  }
+}
