+++ tactical-map/src/index.ts
@@ -1,0 +1,4 @@
+export { TacticalMap } from './tactical-map'
+export * from './types'
+
+console.log(1)
