+++ tactical-map/src/math/vec2.math.ts
