+++ tactical-map/src/types.ts
@@ -1,0 +1,12 @@
+import { IStrategy} from '@cohdex/shared'
+
+export interface IRendererOptions {
+  canvas: HTMLCanvasElement
+  height: number
+  width: number
+}
+
+export interface ITacticalMapOptions {
+  strategy: IStrategy
+  rendererOptions: IRendererOptions
+}
