+++ tactical-map/src/utils/image.util.ts
